requirements.txt
streamlit>=1.24.0
pytz>=2023.3
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
openai>=1.6.1
python-dateutil>=2.8.2